Refactor : api파일 따로 생성해서 연결, 테스트 코드 삭제
