+++ src/pages/index.tsx
@@ -2,13 +2,11 @@ import Link from 'next/link';
 
 import { ContentSection } from '@/components/home/ContentSection';
 import { HeroSection } from '@/components/home/HeroSection';
-import AddWineModal from '@/components/Modal/AddWineModal';
 import { Button } from '@/components/ui/button';
 
 export default function Home() {
   return (
     <div className='mx-auto px-[16px] md:px-[20px] xl:px-0 max-w-[1140px]  min-w-[343px] mt-6 xl:mt-[80px]'>
-      <AddWineModal />
       <HeroSection />
       <main>
         <ContentSection />
